static/test/requirements-test.txt
pytest
pytest-timeout
pytest-html
pytest-metadata
pytest-asyncio
requests
python-dotenv
pyjwt
httpx
respx
fastapi
backend
